app/src/main/java/com/cecs491b/thecookout/transcriptionFiles/requirements.txt
fastapi
uvicorn[standard]
python-multipart
speechmatics-batch
google-generativeai
python-dotenv
httpx